refactor(api): move competence reset error
